Trap focus of overlays which end with form inputs.
Visibility:

Inputs for radio and checkboxes are visually hidden, so check
the label visibility of inputs too when determining whether to
trap focus.

Radios:

When tabbing, the checked radio input of a group of radio inputs
is focused. To change the radio selection / focus the arrow keys
are used. Tabbing again takes focus to the next focuable element
after the radio group. Therefore focus must be trapped before
radio inputs which are not checked (as they do not recieve
focus by tabbing).